Added UI to the Rock Paper Scissors game
Added 3 buttons and 3 paragraph elements to the
index.html.

Got reference for all the elements and used grouping
selector for the buttons. Removed the getHumanChoice
function instead using the id attribute to get the
human choice. Removed the console logs that were
displaying the results instead used the paragraphs
to display the results. Added termination condition
at score 5.